+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [v0.3.0] - 2021-03-19
+
 ### Added
 
 - `validate` cmd that validates tracked files.
@@ -39,7 +41,8 @@
 - Untracking of files.
 - Tracking registry initialization.
 
-[unreleased]: https://github.com/slok/agebox/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/slok/agebox/compare/v0.3.0...HEAD
+[v0.3.0]: https://github.com/slok/agebox/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/slok/agebox/compare/v0.1.1...v0.2.0
 [v0.1.1]: https://github.com/slok/agebox/compare/v0.1.0...v0.1.1
 [v0.1.0]: https://github.com/slok/agebox/releases/tag/v0.1.0